Move logic to handle custom maps out of Story mode
This way we can make it available as a general feature.
