+++ rlbot_gui/story/story-default.json
@@ -21,7 +21,6 @@
                     "humanTeamSize": 1,
                     "opponentBots": ["adversity"],
                     "map": "Simplicity.upk",
-                    "customMap": true,
                     "max_score": "1 Goal",
                     "display": "Custom Map world!"
                 }
