+++ rlbot_gui/story/story_challenge_setup.py
@@ -29,12 +29,8 @@ from rlbot.matchconfig.match_config import (
 )
 from rlbot.setup_manager import SetupManager, RocketLeagueLauncherPreference
 
-from rlbot_gui.match_runner.match_runner import get_fresh_setup_manager
-from rlbot_gui.match_runner.custom_maps import (
-    prepare_custom_map,
-    identify_map_directory,
-    convert_custom_map_to_path
-)
+from rlbot_gui.match_runner.match_runner import get_fresh_setup_manager, setup_match
+
 
 from rlbot_gui import gui as rlbot_gui  # TODO: Need to remove circular import
 
@@ -447,16 +443,7 @@ def manage_game_state(
 
     return calculate_completion(challenge, stats_tracker.stats, results), results
 
-def setup_match(
-    setup_manager: SetupManager, match_config: MatchConfig, launcher_pref: RocketLeagueLauncherPreference
-):
-    """Starts the match and bots"""
-    setup_manager.early_start_seconds = 5
-    setup_manager.connect_to_game(launcher_preference=launcher_pref)
-    setup_manager.load_match_config(match_config)
-    setup_manager.launch_early_start_bot_processes()
-    setup_manager.start_match()
-    setup_manager.launch_bot_processes()
+
 
 
 def run_challenge(
@@ -464,20 +451,7 @@ def run_challenge(
 ) -> Tuple[bool, dict]:
     """Launch the game and keep track of the state"""
     setup_manager = get_fresh_setup_manager()
-
-    if challenge.get("customMap", False):
-        map_file = convert_custom_map_to_path(challenge["map"])
-        rl_directory = identify_map_directory(launcher_pref)
-
-        if not all([map_file, rl_directory]):
-            print("Couldn't load custom map")
-            return False, {}
-
-        with prepare_custom_map(map_file, rl_directory) as game_map:
-            match_config.game_map = game_map
-            setup_match(setup_manager, match_config, launcher_pref)
-    else:
-        setup_match(setup_manager, match_config, launcher_pref)
+    setup_match(setup_manager, match_config, launcher_pref)
 
     setup_manager.game_interface.renderer.clear_screen(RENDERING_GROUP)
     game_results = None
